meval
Module shunting_yard

Implementation of the shunting-yard algorithm for converting an infix expression to an expression in reverse Polish notation (RPN).

See the Wikipedia articles on the shunting-yard algorithm and on reverse Polish notation for more details.

Enums

RPNError

An error produced by the shunting-yard algorightm.

Functions

to_rpn

Converts a tokenized infix expression to reverse Polish notation.